Disable timing tests on darwin
The runtime on darwin uses the wallclock for timeres. Timer tests
running on machines while the clock is adjusted (eg with NTP) can cause
tests to fail. TravisCI VM's seem to be doing this. Disabling timer
related test on OSX for now.

Issue is solved in Go 1.9: https://go-review.googlesource.com/c/35292